mcp/requirements.txt
mcp[cli]>=1.21.2
openai==1.55.3
httpx==0.27.2
python-dotenv==1.0.1
requests==2.32.3
